+++ dash-vue2-alpha/src/main.js
@@ -1,6 +1,6 @@
 import Vue from 'vue'
 import App from './App.vue'
-import router from './router'
+import router from './router/router'
 import vuetify from './plugins/vuetify';
 import {myProjectAuth} from "./backend/backend";
